💄 Proper max width on landing page
Rearraged some of the containers. Reformatted the layout of a couple of the sections. Renamed some of the css class names to improve consistency. Made some judgement calls that will need to be reviewed. Project spotlight had a smaller containerized phone. Well, with the max width, those phones can easily grow and safely stay in their respective areas. So I let em grow.